First step in the implementation of past due date invoices: created a service that computes the due date(s) and outstanding amount(s) based on the Invoice terms; added initial support for multiple payment terms per invoice (e.g. 'pay after 30 days the 50% of the invoice, pay after 60 days the 40%, pay after 90 the 10%'). Part of issue OFBIZ-1530 Modified the getDayEnd util method to be usable in Minilang.
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@604922 13f79535-47bb-0310-9956-ffa450edef68

+++ applications/accounting/src/org/ofbiz/accounting/finaccount/FinAccountServices.java
@@ -201,7 +201,7 @@ public class FinAccountServices {
             }
 
             // set the dates/userlogin
-            inContext.put("thruDate", UtilDateTime.getDayEnd(now, accountValidDays.intValue()));
+            inContext.put("thruDate", UtilDateTime.getDayEnd(now, accountValidDays));
             inContext.put("fromDate", now);
             inContext.put("userLogin", userLogin);
